actually finish moving `metrics` to `unsafe`
