+++ core/js/src/main/scala/cats/effect/IOApp.scala
@@ -16,9 +16,9 @@
 
 package cats.effect
 
-import cats.effect.metrics.{CpuStarvationWarningMetrics, JsCpuStarvationMetrics}
 import cats.effect.std.Console
 import cats.effect.tracing.TracingConstants._
+import cats.effect.unsafe.metrics.{CpuStarvationWarningMetrics, JsCpuStarvationMetrics}
 
 import scala.concurrent.CancellationException
 import scala.concurrent.duration._
