+++ core/js/src/main/scala/cats/effect/unsafe/metrics/JsCpuStarvationMetrics.scala
@@ -14,18 +14,18 @@
  * limitations under the License.
  */
 
-package cats.effect.metrics
+package cats.effect.unsafe.metrics
 
 import cats.effect.IO
 
 import scala.concurrent.duration.FiniteDuration
 
-private[effect] class NativeCpuStarvationMetrics extends CpuStarvationMetrics {
+private[effect] class JsCpuStarvationMetrics extends CpuStarvationMetrics {
   override def incCpuStarvationCount: IO[Unit] = IO.unit
 
   override def recordClockDrift(drift: FiniteDuration): IO[Unit] = IO.unit
 }
 
-private[effect] object NativeCpuStarvationMetrics {
-  private[effect] def apply(): CpuStarvationMetrics = new NativeCpuStarvationMetrics
+private[effect] object JsCpuStarvationMetrics {
+  private[effect] def apply(): CpuStarvationMetrics = new JsCpuStarvationMetrics
 }
