+++ core/jvm/src/main/scala/cats/effect/IOApp.scala
@@ -16,9 +16,9 @@
 
 package cats.effect
 
-import cats.effect.metrics.{CpuStarvationWarningMetrics, JvmCpuStarvationMetrics}
 import cats.effect.std.Console
 import cats.effect.tracing.TracingConstants._
+import cats.effect.unsafe.metrics.{CpuStarvationWarningMetrics, JvmCpuStarvationMetrics}
 import cats.syntax.all._
 
 import scala.concurrent.{blocking, CancellationException, ExecutionContext}
