+++ core/native/src/main/scala/cats/effect/unsafe/metrics/NativeCpuStarvationMetrics.scala
@@ -14,12 +14,18 @@
  * limitations under the License.
  */
 
-package cats.effect.metrics
+package cats.effect.unsafe.metrics
+
+import cats.effect.IO
 
 import scala.concurrent.duration.FiniteDuration
 
-final case class CpuStarvationWarningMetrics(
-    occurrenceTime: FiniteDuration,
-    clockDrift: FiniteDuration,
-    starvationThreshold: Double,
-    starvationInterval: FiniteDuration)
+private[effect] class NativeCpuStarvationMetrics extends CpuStarvationMetrics {
+  override def incCpuStarvationCount: IO[Unit] = IO.unit
+
+  override def recordClockDrift(drift: FiniteDuration): IO[Unit] = IO.unit
+}
+
+private[effect] object NativeCpuStarvationMetrics {
+  private[effect] def apply(): CpuStarvationMetrics = new NativeCpuStarvationMetrics
+}
